Only refresh if ad is visible for aframe
